Crate arc_es_postgres

Re-exports

pub use read_model_store::PostgresReadModelStore;

Modules

read_model_store

Postgres Read Model Store

Structs

PostgresEventStore

Postgres implementation of [EventStore].

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Derive Macros

Deserialize

Serialize